fair: change bias_to_prev_cpu heuristic
The bias_to_prev_cpu heuristic makes a currently
non-sensical check against the last time the task
performed the same check, which makes no sense.
This patch alters that condition to check the
last time the task was descheduled, which can
reasonably affect whether the scheduler should
bias the task back to the same cpu.

Bug: 110464019
Change-Id: I4584db90d9049a2d42077fdb3f88a2c2a3bb5336
Signed-off-by: John Dias <joaodias@google.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>