tcp_westwood : Bump minimum RTT and lower initial RTT duration
Slightly bumping the minimum round-trip-time adds more tolerance
for connections with fluctutating ping to stay connected before
the sender-side TCP stack gives up on the packet and considers
it lost and the connection broken.

Lowering the initial RTT duration works conversely by reducing
tolerance for bad networks so that we avoid getting on bad
connections in the first place.

Co-authored-by: Chinmay Pai <chinmaydpai@gmail.com>
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>
Change-Id: I16ebc432b86acb89c24edecece31e7663e1d7fc7
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>